requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.8.0
anyio==4.3.0
asgiref==3.9.1
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.3.1
cassandra-driver==3.28.0
category-encoders==2.6.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.2.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.7
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
dm-tree==0.1.8
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
Flask==2.0.2
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.4.0
gdown==5.1.0
geomet==0.2.1.post1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googledrivedownloader==0.4
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
imageio==2.34.0
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
install==1.3.5
ipykernel==6.25.2
ipynb-py-convert==0.4.6
ipython==8.15.0
ipywidgets==8.1.1
itsdangerous==2.0.1
jedi==0.19.0
Jinja2==3.0.3
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kaleido==0.2.1
keras==2.13.1
keras-vis==0.4.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lightgbm==4.1.0
lime==0.2.0.1
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.0.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
MiniSom==2.3.2
mlxtend==0.23.1
mpmath==1.3.0
multidict==6.0.5
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.7
outcome==1.3.0.post0
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.1
pivottablejs==0.9.0
platformdirs==3.10.0
plotly==5.17.0
plotly-resampler==0.9.1
pmdarima==2.0.3
prompt-toolkit==3.0.39
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.1.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydotplus==2.0.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.6.1
pyod==1.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytorch-tabnet==4.1.0
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.10.3
rsa==4.9
schemdraw==0.15
scikit-base==0.5.1
scikit-image==0.22.0
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
selenium==4.34.2
six==1.16.0
sktime==0.21.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.31
sqlparse==0.5.3
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tbats==1.1.3
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.23.0
Teras==0.2.0
termcolor==2.3.0
tf-keras-vis==0.8.6
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.2.12
torch==2.2.0
torch_geometric==2.4.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.10.0
trio==0.30.0
trio-websocket==0.12.2
tsdownsample==0.1.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.1.15
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.24.6
vis==0.0.5
wcwidth==0.2.6
websocket-client==1.8.0
Werkzeug==2.0.2
widgetsnbextension==4.0.9
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.0.0
xlrd==2.0.1
xxhash==3.3.0
yarl==1.9.4
yellowbrick==1.5
zipp==3.17.0
